99% complete

+++ src/app/modules/product/product.service.ts
@@ -1,3 +1,5 @@
+import QueryBuilder from "../../builder/QueryBuilder";
+import AppError from "../../errors/AppError";
 import { IProduct } from "./product.interface";
 import { Product } from "./product.schema";
 
@@ -6,8 +8,12 @@ const createSingleProductIntoDB = async (data: IProduct) => {
   return result;
 };
 
-const getAllProductFromDB = async () => {
-  const result = await Product.find();
+const getAllProductFromDB = async (query: Record<string, unknown>) => {
+  const prodctQuery = new QueryBuilder(Product.find(), query)
+    .searchMethod(["title", "category"])
+    .filterMethod()
+    .sortMethod();
+  const result = await prodctQuery.modelQuery;
   return result;
 };
 const getSingleProductFromDB = async (id: string) => {
@@ -35,7 +41,16 @@ const updateSingleProductFromDB = async (
 };
 
 const deleteSingleProductFromDB = async (_id: string) => {
+  const isExist = await Product.findById(_id);
+  if (!isExist) {
+    throw new AppError(404, "Item Not Found to Delete");
+  }
   const result = await Product.findByIdAndDelete(_id);
+  return result;
+};
+
+const getAllCategoryFromDB = async () => {
+  const result = await Product.find();
   return result;
 };
 
@@ -45,4 +60,5 @@ export const productService = {
   updateSingleProductFromDB,
   deleteSingleProductFromDB,
   getSingleProductFromDB,
+  getAllCategoryFromDB,
 };
